requirements.txt
# Core dependencies
# Python 3.10 - 3.13 supported (remove python line as it's not installable via pip)

# OCR and text extraction
pytesseract>=0.3.10
easyocr>=1.7.0
Pillow>=10.0.0

# PDF processing
pdfplumber>=0.9.0
PyPDF2>=3.0.1
PyMuPDF>=1.23.0
pdf2image>=1.16.3
poppler-utils>=0.1.0

# Excel and PowerPoint processing
openpyxl>=3.1.2
python-pptx>=0.6.21

# NLP and PII detection
spacy>=3.7.0
nltk>=3.8.1

# Computer vision for logo detection
opencv-python>=4.8.0
numpy>=1.24.0

# Data processing and analysis
pandas>=2.0.0
scikit-learn>=1.3.0

# Report generation
matplotlib>=3.7.0
graphviz>=0.20.1

# Utility libraries
tqdm>=4.65.0
